requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.5
PyJWT==2.8.0
pyodbc
python-dotenv==1.0.0
Werkzeug==2.3.7
email-validator==2.0.0
click==8.1.7
